String::toUInt8Array() accepts hex value larger than 0xFF.

+++ includes/CxxUtilities/Array.hh
@@ -12,13 +12,29 @@
 
 namespace CxxUtilities {
 
+/**
+ * A class which provides utility functions related to array and vector.
+ */
+
 template<class T>
 class Array: public std::vector<T> {
 public:
+	/** Joins number elements of an array (vector) into a string.
+	 *  @param data array (vector) to be joined.
+	 *  @param dec (for decimal; e.g. 123) or hex (for hexagonal; e.g. 0xAB).
+	 *  @param maxBytesToBeDumped the maximum number of elements to be actually joined (for long array).
+	 *  @return joined string.
+	 */
 	static std::string toString(std::vector<T>& data, std::string mode = "dec", int maxBytesToBeDumped = 8) {
 		return toString(&data, mode, maxBytesToBeDumped);
 	}
 
+	/** Joins number elements of an array (vector) into a string.
+	 *  @param data array (vector) to be joined.
+	 *  @param dec (for decimal; e.g. 123) or hex (for hexagonal; e.g. 0xAB).
+	 *  @param maxBytesToBeDumped the maximum number of elements to be actually joined (for long array).
+	 *  @return joined string.
+	 */
 	static std::string toString(std::vector<T>* data, std::string mode = "dec", int maxBytesToBeDumped = 8) {
 		using namespace std;
 		stringstream ss;
@@ -32,7 +48,7 @@ public:
 			if (mode == "dec") {
 				ss << dec << left << data->at(i);
 			} else if (mode == "hex") {
-				ss << hex << "0x" << setw(2) << setfill('0') << right << data->at(i);
+				ss << hex << "0x" << setw(2) << setfill('0') << right << (uint32_t)data->at(i);
 			} else {
 				ss << data->at(i);
 			}
@@ -53,6 +69,11 @@ public:
 	}
 
 public:
+	/** Splits a string into an array.
+	 * @param str a string to be split.
+	 * @param delimitter a character or a word used as a delimitter.
+	 * @return a newly create array instance.
+	 */
 	static Array<std::string> splitString(std::string str, std::string delimitter) {
 		Array<std::string> result;
 		int n;
@@ -69,6 +90,11 @@ public:
 		return result;
 	}
 
+	/** Splits a string into an array.
+	 * @param str a string to be split.
+	 * @param delimitter a character or a word used as a delimitter.
+	 * @return pointer to a newly created Array instance.
+	 */
 	static Array<std::string>* newBySplittingString(std::string str, std::string delimitter) {
 		Array<std::string>* result=new Array<std::string>;
 		int n;
@@ -85,6 +111,11 @@ public:
 		return result;
 	}
 
+	/** Splits a string into an array.
+	 * @param str a string to be split.
+	 * @param delimitter a character or a word used as a delimitter.
+	 * @return pointer to a newly created Array instance.
+	 */
 	static Array<std::string>* arrayBySplittingString(std::string str, std::string delimitter) {
 		return newBySplittingString(str, delimitter);
 	}
